+++ crates/egui-states/src/client/mod.rs
@@ -3,7 +3,7 @@ mod event;
 pub(crate) mod atomics;
 pub(crate) mod client;
 pub(crate) mod data;
-pub(crate) mod data_multi;
+pub(crate) mod data_take;
 pub(crate) mod image;
 pub(crate) mod messages;
 pub(crate) mod states_creator;
