+++ crates/egui-states/src/lib.rs
@@ -22,8 +22,8 @@ pub use client::{
     atomics::{Atomic, AtomicLock, AtomicLockStatic, AtomicStatic, FallbackLock, UpdateLock},
     client::ClientBuilder,
     client::{Client, ConnectionState},
-    data::{Data, DataTake},
-    data_multi::{DataMulti, DataMultiTake},
+    data::{Data, DataMulti},
+    data_take::{DataMultiTake, DataTake},
     image::ValueImage,
     states_creator::StatesCreator,
     value_map::ValueMap,
