+++ crates/egui-states/src/server/mod.rs
@@ -3,8 +3,8 @@ mod sender;
 mod server_core;
 mod socket_reader;
 
-pub(crate) mod data_multi_server;
 pub(crate) mod data_server;
+pub(crate) mod data_take_server;
 pub(crate) mod image_server;
 pub(crate) mod map_server;
 pub(crate) mod server;
