+++ crates/egui-states/src/server/server.rs
@@ -12,8 +12,8 @@ use crate::data_transport::DataType;
 use crate::event_async::Event;
 use crate::hashing::{NoHashMap, generate_value_id};
 use crate::serialization::{ServerHeader, serialize};
-use crate::server::data_server::{Data, DataTake};
-use crate::server::data_multi_server::{DataMulti, DataMultiTake};
+use crate::server::data_server::{Data, DataMulti};
+use crate::server::data_take_server::{DataMultiTake, DataTake};
 use crate::server::image_server::ValueImage;
 use crate::server::map_server::ValueMap;
 use crate::server::sender::{MessageReceiver, MessageSender};
